pyproject.toml
[project]
name = "fourier-analysis"
version = "0.1.0"
description = "Companion code for 'An Introduction to Fourier Analysis'"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.26",
    "matplotlib>=3.8",
    "scikit-image>=0.22",
    "Pillow>=10.0",
    "scipy>=1.12",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "mypy", "pre-commit"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fourier_analysis"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.mypy]
python_version = "3.12"
strict = true
